frecon: fix logging to /dev/kmsg
frecon is setup to log to /dev/kmsg. Each call to LOG consisted of
several calls to write(): one for the prefix, another for the message,
and a final one to add a trailing newline. The kernel interprets each of
these write() calls as a separate log entry, causing the message to be
broken into contextless fragments. Instead, make a single call to
write().

Include the severity of the log message using the prefix notation
understood by the kernel. The LOG function previously ignored the
severity argument. The log_tsm function converted them to a string and
included them in the message prefix.

Use "frecon" as the syslog tag for the log_tsm function to make it
easier to group messages from frecon. Add pid to messages from log_tsm.

Surround the pid in square brackets instead of round brackets. This is
the standard way syslog expresses it, and consumers of kernel messages
such as journald expect them.

BUG=chromium:929010
TEST=frecon messages in dmsg are correctly formatted
TEST=frecon messages in journalctl have correct _PID and
SYSLOG_IDENTIFIER values

Change-Id: I3fa80d51759daf1cda51b3a4adec46f9bd26272f
Reviewed-on: https://chromium-review.googlesource.com/1456564
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Christopher Morin <cmtm@google.com>
Reviewed-by: Dominik Behr <dbehr@chromium.org>