input: Use KEY_FORWARD for VT2 when present
Developers are used to use the T2 key for many years. And they are
getting confused when they realize that on the older style keyboards
that still has the "FORWARD" key, they can not longer use T2. Lets
enable FORWARD also to swtich to VT2, so that users are not taken by
surprise.

BUG=b:156019833
TEST=None

Change-Id: Ibe04a2b98fb49680dc80f66fac1436caf19fb444
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/frecon/+/2199938
Reviewed-by: Dominik Behr <dbehr@chromium.org>
Tested-by: Rajat Jain <rajatja@google.com>
Commit-Queue: Rajat Jain <rajatja@google.com>